remove draft file
